+++ experiments/notebooks/visualizations/__init__.py
@@ -674,7 +674,7 @@ def fig_add_stage_markers(df, column, fig, secondary_y=None, parameters=paramete
     system_dates = [
         ("Beacon Chain", datetime.strptime("Dec-01-2020", "%b-%d-%Y")),
         # ("Today", parameters["date_start"][0]),
-        ("EIP1559", parameters["date_eip1559"][0]),
+        ("EIP-1559", parameters["date_eip1559"][0]),
         ("Proof of Stake", parameters["date_pos"][0]),
     ]